Add files via upload
This code is similiar to having data sent to a dump file after inputs on a console in my opinion. What is actually identifies as is a program taking a list of integers and chars for example:

618 phi
777 boeing
737 boeingoriginal

and seperating the integers into arrays and then placing the chars next to it also in the same array. Afterwards it places this list into an output file. Sorta gives you an understanding of how consoles and dump files work